* Refactory manager for multiprocess excution (multi-parallel execution). * Put back old version of maximality algorithm for bechmarking with version-2 maximality * Repair the first version of maximality algorithm (considering the indifference decisions)
